gan modifications

+++ discriminator/ImageDataset.py
@@ -1,5 +1,3 @@
-import torch
-import numpy as np
 from torch.utils.data import Dataset
 
 
@@ -20,8 +18,8 @@ class ImageDataset(Dataset):
 
         self.img_list = []
         for i in range(len(self.reals)):
-            real = self.transform(self.reals[i].astype(np.uint8))
-            fake = self.transform(self.fakes[i].astype(np.uint8))
+            real = self.transform(self.reals[i])
+            fake = self.transform(self.fakes[i])
             self.img_list.append([real, fake])
 
     def __len__(self):
